fix for #105 , #102 and #98
Fixes #105 , Fixes #102 , Fixes #98

+++ comic_dl/globalFunctions.py
@@ -16,8 +16,6 @@ import img2pdf
 
 class GlobalFunctions(object):
     def page_downloader(self, manga_url, **kwargs):
-        # manga_url = str(manga_url).decode('unicode_escape').encode('utf-8')
-        # print("Manga URL : {0}".format(manga_url))
         headers = kwargs.get("headers")
         if not headers:
             headers = {
@@ -104,8 +102,9 @@ class GlobalFunctions(object):
             print("\n")
             print('{:^80}'.format('====================================================================='))
             # print('{:^80}'.format("%s - %s" % (anime_name, episode_number)))
-            print('{:^80}'.format("Manga Name : %s" % anime_name))
-            print('{:^80}'.format("Chapter Number - %s" % episode_number))
+            print('{:^80}'.format("Manga Name : {0}".format(anime_name)))
+            print('{:^80}'.format("Chapter Number - {0}".format(episode_number)))
+            print('{:^80}'.format("Total Chapters - {0}".format(kwargs.get('total_chapters'))))
             print('{:^80}'.format('=====================================================================\n'))
 
     def conversion(self, directory_path, conversion, delete_files, comic_name, chapter_number):
@@ -124,9 +123,14 @@ class GlobalFunctions(object):
                                                                   str((x.split('.')[0])).split(os.sep)[-1]))]
             pdf_file_name = str(converted_file_directory) + "{0} - Ch {1}.pdf".format(comic_name, chapter_number)
             try:
-                with open(pdf_file_name, "wb") as f:
-                    f.write(img2pdf.convert(im_files))
-                    print("Converted the file to pdf...")
+                # This block is same as the one in the "cbz" conversion section. Check that one.
+                if os.path.isfile(pdf_file_name):
+                    print('[Comic-dl] CBZ File Exist! Skipping : {0}\n'.format(pdf_file_name))
+                    pass
+                else:
+                    with open(pdf_file_name, "wb") as f:
+                        f.write(img2pdf.convert(im_files))
+                        print("Converted the file to pdf...")
             except Exception as FileWriteError:
                 print("Couldn't write the pdf file...")
                 print(FileWriteError)
@@ -137,15 +141,23 @@ class GlobalFunctions(object):
         elif str(conversion).lower().strip() in ['cbz']:
 
             cbz_file_name = str(converted_file_directory) + "{0} - Ch {1}".format(comic_name, chapter_number)
+            print("CBZ File : {0}".format(cbz_file_name))
 
             try:
-                shutil.make_archive(cbz_file_name, 'zip', directory_path, directory_path)
-                os.rename(str(cbz_file_name) + ".zip", (str(cbz_file_name)+".zip").replace(".zip", ".cbz"))
+                """If the .cbz file exists, we don't need to make it again. If we do make it again, it'll make the 
+                .zip file and will hit and exception about file existing already. This raised #105.
+                So, to fix the #105, we'll add this check and make things work just fine."""
+                if os.path.isfile(str(cbz_file_name) + ".cbz"):
+                    print('[Comic-dl] CBZ File Exist! Skipping : {0}\n'.format(cbz_file_name))
+                    pass
+                else:
+                    shutil.make_archive(cbz_file_name, 'zip', directory_path, directory_path)
+                    os.rename(str(cbz_file_name) + ".zip", (str(cbz_file_name)+".zip").replace(".zip", ".cbz"))
             except Exception as CBZError:
                 print("Couldn't write the cbz file...")
                 print(CBZError)
                 # Let's not delete the files if the conversion failed...
-                keep_files = "False"
+                keep_files = "True"
                 pass
 
         elif str(conversion) == "None":
@@ -166,7 +178,7 @@ class GlobalFunctions(object):
     
     def saveNewRange(self, comicUrl, nextChapterIndex):
         # @dsanchezseco
-        #edit config.json to update nextChapter value
+        # edit config.json to update nextChapter value
         data = json.load(open('config.json'))
         data["comics"][comicUrl]["next"] = data["comics"][comicUrl]["next"] + nextChapterIndex
         json.dump(data, open('config.json', 'w'), indent=4)
